+++ Fuji/Source/Util.cpp
@@ -99,7 +99,7 @@ MF_API MFEndian MFSystem_GetPlatformEndian(int platform)
 // Return only the last portion of the module name (exclude the path)
 char *ModuleName(char *pSourceFileName)
 {
-	int x = MFString_Length(pSourceFileName)-1;
+	int x = (int)MFString_Length(pSourceFileName)-1;
 
 	while(x >= 0 && pSourceFileName[x] != '/' && pSourceFileName[x] != '\\')
 	{
